Work-in-progress log object in qt5gui.
For now just forwarding to std::cout. This will become a proper GUI log
widget in the future.
